Avoid circular modules.
